refactor: create nic api classes
